fix: 생산기획 추천 API SQL 타입 캐스팅 오류 수정
$1 파라미터가 문자열 연결($1 || ' days')과 숫자 나눗셈에 동시 사용되어
PostgreSQL이 text 타입으로 추론, numeric / text 오류 발생.
$1::numeric 명시적 캐스팅으로 해결.

+++ server/src/modules/production/production.repository.ts
@@ -228,8 +228,8 @@ class ProductionRepository extends BaseRepository<ProductionPlan> {
             ELSE 'SA'
           END AS product_season,
           COALESCE(SUM(s.qty), 0)::int AS total_sold,
-          ROUND(COALESCE(SUM(s.qty), 0)::numeric / $1, 2) AS avg_daily_sales,
-          ROUND(COALESCE(SUM(s.qty), 0)::numeric / $1 * 30)::int AS predicted_30d
+          ROUND(COALESCE(SUM(s.qty), 0)::numeric / $1::numeric, 2) AS avg_daily_sales,
+          ROUND(COALESCE(SUM(s.qty), 0)::numeric / $1::numeric * 30)::int AS predicted_30d
         FROM products p
         JOIN product_variants pv ON p.product_code = pv.product_code
         JOIN sales s ON pv.variant_id = s.variant_id
